requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
python-dotenv
roboflow
pillow
inference-sdk
aiofiles      
